+++ src/main/java/com/goat/server/mypage/dto/request/GoalRequest.java
@@ -1,4 +1,4 @@
-package com.goat.server.mypage.dto;
+package com.goat.server.mypage.dto.request;
 
 public record GoalRequest(
         String goal
