+++ src/main/java/com/goat/server/mypage/presentation/MypageController.java
@@ -2,8 +2,8 @@ package com.goat.server.mypage.presentation;
 
 import com.goat.server.global.dto.ResponseTemplate;
 import com.goat.server.mypage.application.MypageService;
-import com.goat.server.mypage.dto.GoalRequest;
-import com.goat.server.mypage.dto.UserMajorResponse;
+import com.goat.server.mypage.dto.request.GoalRequest;
+import com.goat.server.mypage.dto.response.UserMajorResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,6 @@ import org.springframework.web.bind.annotation.*;
 public class MypageController {
 
     private final MypageService mypageService;
-
     @Operation(summary = "마이페이지 정보 보기", description = "마이페이지 정보 보기")
     @GetMapping("/info/{userId}")
     public ResponseEntity<ResponseTemplate<Object>> getUserDetails(@PathVariable Long userId) {
